Move tablet page to its own app
