+++ fars/booking/views.py
@@ -209,49 +209,6 @@ class BookingView(View):
             return False, _("Only the user or group that made the booking may unbook it")
         return True, ''
 
-
-class TabletView(View):
-    template = 'tablet.html'
-    context = {}
-
-    def dispatch(self, request, bookable):
-        bookable_obj = get_object_or_404(Bookable, id_str=bookable)
-        if not bookable_obj.public and not request.user.is_authenticated:
-            return redirect('{}?next={}'.format(reverse('login'), request.path_info))
-        self.context['bookable'] = bookable_obj
-        self.context['errors'] = 0
-        return super().dispatch(request, bookable)
-
-    def get(self, request, bookable):
-        now = pytz.timezone(TIME_ZONE).localize(datetime.now())
-        booking = Booking()
-        booking.bookable = self.context['bookable']
-        # self.context['bookform'] = BookingForm(instance=booking)
-        return render(request, self.template, context=self.context)
-
-
-    def post(self, request, bookable):
-        username = request.POST.get('username')
-        pw = request.POST.get('password')
-        user = authenticate(username=username, password=pw)
-        postdata = request.POST.copy()
-        try:
-            postdata['user'] = user.id
-        except AttributeError:
-            postdata['user'] = -1
-            self.context['errors'] = 1
-            self.context['credential_error'] = 1
-        form = BookingForm(postdata, instance=Booking())
-        if form.is_valid():
-            form.save()
-            return redirect('tablet', bookable=bookable)
-        self.context['errors'] = 1
-        status = 400
-
-        self.context['bookform'] = form
-        return render(request, self.template, context=self.context, status=status)
-
-
 # Returns whether user is admin for given bookable
 def _is_admin(user, bookable):
     return user.is_superuser or user.groups.filter(name=bookable.admin_group_name).exists()
