+++ fars/tabletpage/static/css/tablet.css
