+++ fars/tabletpage/static/js/tablet.js
