+++ fars/tabletpage/templates/tablet.html
