+++ fars/tabletpage/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from . import views
+
+urlpatterns = [
+   path('<slug:bookable>', views.TabletView.as_view(), name="tablet"),
+]
